requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pydantic
pypdf
faiss-cpu
sentence-transformers
openai
ollama
loguru
click
requests
